Fix quitting with Ctrl+C
In order for Ctrl + C to stop the Docker image, we have to prevent
Docker from introducing an intermediate `sh` process that will
capture the SIGINT signal. To do so, we have to use the script as
our ENTRYPOINT.

Since the entrypoint script is another Bash shell, we need to use
`exec` when starting the Python HTTP server so it will replace the
Bash shell process, and become the parent process of the functions
run by Maven.